+++ config/test.exs
@@ -56,4 +56,17 @@ config :oli, OliWeb.Endpoint,
   server: false
 
 # Print only warnings and errors during test
-config :logger, level: :warn, truncate: :infinity
+config :logger, level: :warn
+
+truncate =
+  System.get_env("LOGGER_TRUNCATE", "8192")
+  |> String.downcase()
+  |> case do
+    "infinity" ->
+      :infinity
+
+    val ->
+      String.to_integer(val)
+  end
+
+config :logger, truncate: truncate