+++ lib/oli_web/live/delivery/remix/remove_modal.ex
@@ -6,9 +6,9 @@ defmodule OliWeb.Delivery.Remix.RemoveModal do
 
   alias Oli.Delivery.Hierarchy.HierarchyNode
 
-  def render(%{node: %HierarchyNode{slug: slug, revision: revision}} = assigns) do
+  def render(%{node: %HierarchyNode{uuid: uuid, revision: revision}} = assigns) do
     ~L"""
-    <div class="modal fade show" id="delete_<%= slug %>" tabindex="-1" role="dialog" aria-hidden="true" phx-hook="ModalLaunch">
+    <div class="modal fade show" id="delete_<%= uuid %>" tabindex="-1" role="dialog" aria-hidden="true" phx-hook="ModalLaunch">
       <div class="modal-dialog" role="document">
         <div class="modal-content">
             <div class="modal-header">
@@ -25,8 +25,7 @@ defmodule OliWeb.Delivery.Remix.RemoveModal do
               <button
                 phx-click="RemoveModal.remove"
                 phx-key="enter"
-                phx-value-slug="<%= slug %>"
-                onclick="$('#delete_<%= slug %>').modal('hide')"
+                phx-value-uuid="<%= uuid %>"
                 class="btn btn-danger">
                 Remove
               </button>